[PM-34857] Remove PoliciesInAcceptedState feature flag (#21761)
Keeping both old and new state and response properties
for now, state migration will be a separate PR.
